change color setting and fixed next link issue
