+++ package.json
@@ -24,7 +24,7 @@
     "eslint": "8.49.0",
     "eslint-config-next": "13.5.2",
     "framer-motion": "^10.16.4",
-    "next": "13.5.2",
+    "next": "^14.0.5-canary.40",
     "react": "18.2.0",
     "react-colorful": "^5.6.1",
     "react-dom": "18.2.0",
