+++ styles/home.module.scss
@@ -8,6 +8,10 @@
 .maintext {
     @include flexCenter;
     height: 100vh;
+    width: 100%;
+    position: fixed;
+    top: 0;
+    z-index: -99;
 
     h1 {
         text-align: center;
@@ -22,7 +26,6 @@
         position: absolute;
         font-size: 8em;
         line-height: 1em;
-        color: #3bff58;
         mix-blend-mode:screen;
         font-weight: 900;
     }
@@ -89,6 +92,11 @@
     }
 }
 
+
+.donutContainer {
+    width: 100%;
+    height: 100%;
+}
 
 .colorPicker {
     position: fixed;
